background and techImage moved to ui
