+++ src/ui/TechImage.jsx
